user/requirements.txt
fastapi
motor
passlib
python-dotenv
pyjwt
uvicorn